+++ src/test/java/com/wireguard/api/peer/PeerControllerTest.java
@@ -3,6 +3,7 @@ package com.wireguard.api.peer;
 import com.wireguard.api.ResourceNotFoundException;
 import com.wireguard.api.converters.WgPeerDTOFromWgPeerConverter;
 import com.wireguard.api.dto.PageDTO;
+import com.wireguard.api.peer.controller.PeerController;
 import com.wireguard.external.network.NoFreeIpException;
 import com.wireguard.external.network.Subnet;
 import com.wireguard.external.network.SubnetV6;
